Add resource parameter to PAR
As recommended by the OpenID4VCI draft 14, use RFC8707 resource parameter when requesting issuance using scope parameter.

The exact phrasing in the draft reads "If the Credential Issuer metadata contains an authorization_servers property, it is RECOMMENDED to use a resource parameter". However, this information is lost in the TO->domain conversion, so it's not taken into account.